+++ packages/components/pie-button/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.47.6
+
+### Patch Changes
+
+- [Changed] - Update `element-internals-polyfill` to v1.3.11 ([#1452](https://github.com/justeattakeaway/pie/pull/1452)) by [@xander-marjoram](https://github.com/xander-marjoram)
+
 ## 0.47.5
 
 ### Patch Changes